requirements.txt
numpy>=1.20.2,<1.21
django>=3.2,<3.3
psycopg2-binary == 2.8.6
flake8>=3.9.1,<3.10.0
djangorestframework>=3.12.4,<3.13
